[JT-B-J-029] The converter has been separated and the tests for it have been moved to a separate package.
