+++ auth/src/main/java/com/jointAuth/converter/UserConverter.java
@@ -1,0 +1,17 @@
+package com.jointAuth.converter;
+
+import com.jointAuth.model.User;
+import com.jointAuth.model.UserDTO;
+
+public class UserConverter {
+    public static UserDTO convertToDto(User user) {
+        UserDTO dto = new UserDTO();
+        dto.setId(user.getId());
+        dto.setFirstName(user.getFirstName());
+        dto.setLastName(user.getLastName());
+        dto.setEmail(user.getEmail());
+        dto.setRegistrationDate(user.getRegistrationDate());
+        dto.setLastLogin(user.getLastLogin());
+        return dto;
+    }
+}
